requirements.txt
datetime
python-dotenv==1.0.1
pandas==2.2.3
pdfplumber==0.11.5
dataclasses
typing
llama-index==0.12.8
llama-index-core==0.12.8
llama-index-embeddings-huggingface==0.5.0
huggingface-hub==0.23.2
openai==1.58.1
tenacity==8.2.2
